Route Switching on isAuth
There is a little bit problem when we logout then signIn Again, it do not sign in because there is no values in email, password, name states. But why? how they are removedwhen logout. It may be happended due to when we logout, it re-render to login screen, there default state has null values, but we setup values in store, store has also no values after logout